reuse linker-dapp api code
