+++ packages/@dcl/sdk-commands/src/linker-dapp/api.ts
@@ -1,17 +1,20 @@
-import { createServerComponent } from '@well-known-components/http-server'
+import { Router, createServerComponent } from '@well-known-components/http-server'
 import { Lifecycle } from '@well-known-components/interfaces'
 import { createRecordConfigComponent } from '@well-known-components/env-config-provider'
 import { createConsoleLogComponent } from '@well-known-components/logger'
 import { resolve } from 'path'
 import { ChainId } from '@dcl/schemas'
 import open from 'open'
-import { IFuture } from 'fp-future'
 
-import { getPort } from '../../../logic/get-free-port'
-import { CliComponents } from '../../../components'
-import { setRoutes } from './routes'
-import { CreateQuest, QuestLinkerActionType } from '../types'
+import { getPort } from '../logic/get-free-port'
+import { CliComponents } from '../components'
 
+export interface LinkerdAppOptions {
+  openBrowser: boolean
+  linkerPort?: number
+  isHttps: boolean
+  uri: string
+}
 export interface LinkerResponse {
   address: string
   signature: string
@@ -20,16 +23,12 @@ export interface LinkerResponse {
 
 export async function runLinkerApp(
   cliComponents: Pick<CliComponents, 'fs' | 'logger' | 'fetch' | 'config'>,
-  awaitResponse: IFuture<void>,
-  port: number,
-  info: { messageToSign: string; extraData?: { questName?: string; questId?: string; createQuest?: CreateQuest } },
-  actionType: QuestLinkerActionType,
-  { isHttps, openBrowser }: { isHttps: boolean; openBrowser: boolean },
-  deployCallback: (linkerResponse: LinkerResponse) => Promise<void>
+  router: Router<object>,
+  { isHttps, openBrowser, uri, linkerPort }: LinkerdAppOptions
 ) {
-  const resolvedPort = await getPort(port)
+  const resolvedPort = await getPort(linkerPort!)
   const protocol = isHttps ? 'https' : 'http'
-  const url = `${protocol}://localhost:${resolvedPort}/quests`
+  const url = `${protocol}://localhost:${resolvedPort}${uri}`
   const program = await Lifecycle.run({
     async initComponents() {
       const config = createRecordConfigComponent({
@@ -47,36 +46,29 @@ export async function runLinkerApp(
     },
     async main(program) {
       const { components, startComponents } = program
-      const { router } = setRoutes(components, awaitResponse, info, actionType, deployCallback)
       components.server.setContext(components)
       components.server.use(router.allowedMethods())
       components.server.use(router.middleware())
 
       await startComponents()
-      if (openBrowser) {
-        await browse(components, url, '')
-      } else {
-        components.logger.info('You need to sign the quest before executing the command:')
-        components.logger.info('')
-        components.logger.info(`Navigate to:\n ${url}`)
-      }
+      if (openBrowser) await browse(components, url)
     }
   })
   return { program }
 }
 
-async function browse({ logger }: Pick<CliComponents, 'logger'>, url: string, params: string) {
-  logger.info('You need to sign the quest before executing the command:')
+async function browse({ logger }: Pick<CliComponents, 'logger'>, url: string) {
+  logger.info('You need to sign the content before the deployment:')
 
   setTimeout(async () => {
     try {
-      await open(`${url}?${params}`)
+      await open(`${url}`)
     } catch (e) {
-      logger.error(`Unable to open browser automatically. Please manually navigate to:\n  ${url}?${params}`)
+      logger.error(`Unable to open browser automatically. Please manually navigate to:\n  ${url}`)
     }
   }, 5000)
 
-  logger.info(`Signing app ready at ${url}?${params}`)
+  logger.info(`Signing app ready at ${url}`)
 }
 
 async function getCredentials({ fs }: Pick<CliComponents, 'fs' | 'logger'>) {
